book: new description of android-example
