+++ android-example/.cargo/config
@@ -1,15 +1,7 @@
 [target.aarch64-linux-android]
-linker = "aarch64-linux-android-gcc"
+linker = "aarch64-linux-android21-clang++"
 runner = "./run-on-adroid.sh"
 
 [target.arm-linux-androideabi]
-linker = "arm-linux-androideabi-gcc"
-runner = "./run-on-adroid.sh"
-
-[target.i686-linux-android]
-linker = "i686-linux-android-gcc"
-runner = "./run-on-adroid.sh"
-
-[target.x86_64-linux-android]
-linker = "x86_64-linux-android-gcc"
+linker = "armv7a-linux-androideabi21-clang++"
 runner = "./run-on-adroid.sh"
