+++ android-tests/.gitignore
@@ -1,0 +1,13 @@
+.idea/
+*.iml
+.gradle
+/local.properties
+.DS_Store
+/build
+/captures
+.externalNativeBuild
+target/
+app/src/main/libs/
+app/src/main/java/net/akaame/myapplication/Session.java
+src/java_glue.rs
+Cargo.lock
