+++ android-tests/app/src/androidTest/java/net/akaame/myapplication/ExampleInstrumentedTest.java
@@ -3,9 +3,7 @@ package net.akaame.myapplication;
 import android.content.Context;
 import android.support.test.InstrumentationRegistry;
 import android.support.test.runner.AndroidJUnit4;
-import android.util.Log;
 
-import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 
@@ -18,23 +16,11 @@ import static org.junit.Assert.*;
  */
 @RunWith(AndroidJUnit4.class)
 public class ExampleInstrumentedTest {
-    private static final String TAG = ExampleInstrumentedTest.class.getSimpleName();
-
-    @Before
-    public void setUp() throws Exception {
-        Log.d(TAG, "setUp");
-        System.loadLibrary("mobcore");
-    }
     @Test
     public void useAppContext() throws Exception {
         // Context of the app under test.
         Context appContext = InstrumentationRegistry.getTargetContext();
 
         assertEquals("net.akaame.myapplication", appContext.getPackageName());
-    }
-    @Test
-    public void testSession() {
-        Session session = new Session();
-        assertEquals(5, session.add_and1(2));
     }
 }
