+++ android-tests/app/src/main/res/values/strings.xml
@@ -1,0 +1,5 @@
+<resources>
+    <string name="app_name">My Application</string>
+    <string name="action_settings">Settings</string>
+    <string name="rust_add1_text">We can do math in Rust and get %1$d</string>
+</resources>
