+++ android-tests/build.gradle
@@ -1,0 +1,35 @@
+// Top-level build file where you can add configuration options common to all sub-projects/modules.
+
+buildscript {
+    repositories {
+        jcenter()
+        google()
+    }
+    dependencies {
+        classpath 'com.android.tools.build:gradle:3.3.2'
+        classpath "org.codehaus.groovy:groovy-json:2.5.6"
+
+        // NOTE: Do not place your application dependencies here; they belong
+        // in the individual module build.gradle files
+    }
+}
+
+allprojects {
+    repositories {
+        jcenter()
+        maven {
+            url "https://maven.google.com"
+        }
+    }
+}
+
+task clean(type: Delete) {
+    delete rootProject.buildDir, 'app/src/main/libs'
+
+    // Optionally run cargo clean - this will result in huge rebuild times
+    // as cargo will redownload and recompile all dependencies
+    /*exec {
+        workingDir '.'
+        commandLine 'cargo', 'clean'
+    }*/
+}
