+++ android-tests/gradle/wrapper/gradle-wrapper.properties
@@ -1,0 +1,6 @@
+#Fri Apr 20 19:57:15 CEST 2018
+distributionBase=GRADLE_USER_HOME
+distributionPath=wrapper/dists
+zipStoreBase=GRADLE_USER_HOME
+zipStorePath=wrapper/dists
+distributionUrl=https\://services.gradle.org/distributions/gradle-4.10.1-all.zip
