+++ android-tests/src/android_c_headers.rs
@@ -1,0 +1,3 @@
+#![allow(non_upper_case_globals, dead_code, non_camel_case_types, improper_ctypes, non_snake_case)]
+
+include!(concat!(env!("OUT_DIR"), "/android_c_headers.rs"));
