+++ android-tests/src/java_glue.rs.in
@@ -1,11 +1,8 @@
-use crate::Session;
-use jni_sys::*;
+use crate::{android_c_headers::*, Session};
 
-// ANCHOR: api
 foreigner_class!(class Session {
     self_type Session;
     constructor Session::new() -> Session;
     fn Session::add_and1(&self, val: i32) -> i32;
     fn Session::greet(to: &str) -> String;
 });
-// ANCHOR_END: api
