+++ android-tests/src/lib.rs
@@ -1,8 +1,16 @@
-use log::info;
+#[cfg(target_os = "android")]
+extern crate android_logger;
+#[macro_use]
+extern crate log;
+extern crate log_panics;
+
+#[cfg(target_os = "android")]
+mod android_c_headers;
+#[cfg(target_os = "android")]
 mod java_glue;
+#[cfg(target_os = "android")]
 pub use crate::java_glue::*;
 
-// ANCHOR: rust_code
 struct Session {
     a: i32,
 }
@@ -29,4 +37,3 @@ impl Session {
         format!("Hello {} ✋\nIt's a pleasure to meet you!", to)
     }
 }
-// ANCHOR_END: rust_code
