+++ book/src/SUMMARY.md
@@ -2,4 +2,6 @@
 
 [About](./about.md)
 - [Getting Started](./getting-started.md)
-  - [C++](./cpp.md)
+  - [C++](./cpp-example.md)
+  - [Java/Android](./java-android-example.md)
+  - [Java/Other](./java-other.md)
