+++ book/src/cpp-example.md
