+++ workspace.Cargo.toml
@@ -1,5 +1,5 @@
 [workspace]
-members = ["macroslib", "jni_tests", "c++_tests", "android-example",
+members = ["macroslib", "jni_tests", "c++_tests", "android-example", "android-tests",
 "debug-util", "cpp-example/rust-part"]
 
 [profile.release]
